Big merge up to current windows release (0.61b2) - part 2
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@185 37977431-3df6-0310-b722-df95706aa16b

+++ rts/WakeProjectile.h
@@ -15,7 +15,7 @@ public:
 	float alpha;
 	float alphaFalloff;
 	float alphaAdd;
-	float alphaAddTime;
+	int alphaAddTime;
 	float size;
 	float sizeExpansion;
 	float rotation;
